# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POS System API for Egyptian retail shop
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List all units
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    patch:
      operationId: updateUnit
      tags: [units]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted

  # Products
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppliersPage"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Supplier profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierProfile"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
  /suppliers/{id}/payments:
    get:
      operationId: getSupplierPayments
      tags: [suppliers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Supplier payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupplierPayment"
    post:
      operationId: addSupplierPayment
      tags: [suppliers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierPaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPayment"

  # Quotations
...
components:
  schemas:
...
    Unit:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
    UnitInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        nameEn: { type: string }
    UnitUpdate:
      type: object
      properties:
        name: { type: string }
        nameEn: { type: string }

    # Product
...
    Supplier:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        openingBalance: { type: number }
        totalDebt: { type: number }
        isActive: { type: boolean }
        createdAt: { type: string }
    SupplierInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        contactPerson: { type: string }
        notes: { type: string }
        openingBalance: { type: number }
        isActive: { type: boolean }
    SupplierUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        contactPerson: { type: string }
        notes: { type: string }
        openingBalance: { type: number }
        isActive: { type: boolean }
    SupplierProfile:
      type: object
      required: [supplier, invoices, totalDebt, totalPaid]
      properties:
        supplier:
          $ref: "#/components/schemas/Supplier"
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseInvoice"
        totalDebt: { type: number }
        totalPaid: { type: number }
    SuppliersPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Supplier"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    SupplierPayment:
      type: object
      required: [id, supplierId, amount, date]
      properties:
        id: { type: integer }
        supplierId: { type: integer }
        supplierName: { type: ["string", "null"] }
        amount: { type: number }
        date: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        createdBy: { type: ["string", "null"] }
    SupplierPaymentInput:
      type: object
      required: [amount, date]
      properties:
        amount: { type: number }
        date: { type: string }
        notes: { type: string }

    # Invoice Item (shared)
...
    PurchaseInvoice:
      type: object
      required: [id, serial, status, total, createdAt]
      properties:
        id: { type: integer }
        serial: { type: string }
        status: { type: string }
        paymentType: { type: string }
        supplierId: { type: ["integer", "null"] }
        supplierName: { type: ["string", "null"] }
        subtotal: { type: number }
        total: { type: number }
        paidAmount: { type: number }
        remainingAmount: { type: number }
        notes: { type: ["string", "null"] }
        invoiceDate: { type: ["string", "null"] }
        createdAt: { type: string }
        createdBy: { type: ["string", "null"] }
...